+++ apig_wsgi.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+from base64 import b64encode
 import sys
 from io import BytesIO
 
@@ -14,14 +15,17 @@ __version__ = '1.0.0'
 __all__ = ('make_lambda_handler',)
 
 
-def make_lambda_handler(wsgi_app):
+def make_lambda_handler(wsgi_app, binary_media_types=None):
     """
     Turn a WSGI app callable into a Lambda handler function suitable for
     running on API Gateway.
     """
+    if binary_media_types is None:
+        binary_media_types = []
+
     def handler(event, context):
         environ = get_environ(event)
-        response = Response()
+        response = Response(binary_media_types=binary_media_types)
         result = wsgi_app(environ, response.start_response)
         response.consume(result)
         return response.as_apig_response()
@@ -70,10 +74,15 @@ def get_environ(event):
 
 
 class Response(object):
-    def __init__(self):
+    def __init__(self, binary_media_types):
         self.status_code = '500'
         self.headers = []
         self.body = BytesIO()
+
+        self.binary_media_types = set(binary_media_types)
+        if '*/*' in self.binary_media_types:
+            # everything is binary
+            self.binary_media_types = None
 
     def start_response(self, status, response_headers, exc_info=None):
         self.status_code = status.split()[0]
@@ -90,8 +99,28 @@ class Response(object):
                 result.close()
 
     def as_apig_response(self):
-        return {
+        response = {
             'statusCode': self.status_code,
             'headers': dict(self.headers),
-            'body': self.body.getvalue().decode('utf-8'),
         }
+
+        content_type = self._get_content_type()
+
+        if self._is_binary_media(content_type):
+            response['isBase64Encoded'] = True
+            response['body'] = b64encode(self.body.getvalue()).decode('utf-8')
+        else:
+            response['body'] = self.body.getvalue().decode('utf-8')
+
+        return response
+
+    def _get_content_type(self):
+        content_type_headers = [v for k, v in self.headers if k.lower() == 'content-type']
+        if len(content_type_headers):
+            return content_type_headers[-1]
+        return None
+
+    def _is_binary_media(self, content_type):
+        if self.binary_media_types is None:
+            return True
+        return content_type in self.binary_media_types