+++ HISTORY.rst
@@ -6,6 +6,9 @@ Pending Release
 
 .. Insert new release notes below this line
 
+* Support binary data for responses with `Content-Type` headers that aren't
+  `application/json` or start with `text/`.
+
 1.0.0 (2018-03-08)
 ------------------